WIP: Test passed, but no way to persist that Trie!!
* Some clean up
